Fix invalid class name in CommentController.php Fix invalid class namespace in BusinessUnit.php Move some munkireport\models into the Compatibility namespace. Move most shim functions into the compatibility dir Update composer.json to use new shim directory Reconcile phpunit.xml with latest Laravel template. Futile struggle with log configuration in unit test environment
