+++ .env.testing
@@ -3,4 +3,5 @@ DB_CONNECTION=sqlite  # Was: CONNECTION_DRIVER
 CONNECTION_DATABASE=:memory:
 DB_DATABASE=:memory:   # Was: CONNECTION_DATABASE
 APP_KEY=base64:ZZxdn6dbVTL7cpzcmALmsSG0uwa+b07KilEHDweqXyQ=
+LOG_CHANNEL=stack
