+++ app/Hash.php
@@ -7,6 +7,7 @@ use Illuminate\Database\Eloquent\Model;
 class Hash extends Model
 {
     protected $table = 'hash';
+    public $timestamps = false;
 
     protected $fillable = [
         'serial_number',
