+++ compatibility/MachineGroup.php
@@ -2,12 +2,15 @@
 
 namespace Compatibility;
 
+use Compatibility\Kiss\Contracts\LegacyMachineGroup;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Collection;
 use Webpatser\Uuid\Uuid;
 
 /**
  * MachineGroup model for v5 compatibility.
+ *
+ * It is not possible to implement compatibility with the ::all() method of the legacy model
  */
 class MachineGroup extends Model
 {
@@ -103,4 +106,13 @@ class MachineGroup extends Model
     }
 
     //// SCOPES
+    public function get_max_groupid(): int
+    {
+        // TODO: Implement get_max_groupid() method.
+    }
+
+    public function get_group_ids(): array
+    {
+        // TODO: Implement get_group_ids() method.
+    }
 }
