--- shims/Machine_model.php
@@ -1,8 +1,0 @@
-<?php
-
-/**
- * This class alias provided for backwards compatibility when 3rd party modules access machine via
- * \Machine_model without the namespace.
- */
-//class_alias('munkireport\models\Machine_model', '\Machine_model');
-class_alias('munkireport\models\Machine_model', '\App\Machine');
