--- shims/Reportdata_model.php
@@ -1,7 +1,0 @@
-<?php
-
-/**
- * This class alias provided for backwards compatibility when 3rd party modules access reportdata via
- * \Reportdata_model without the namespace.
- */
-class_alias('App\ReportData', '\Reportdata_model');
